using thymeleaf extras for sec: attributes and code review comments addressed
