+++ ideaforce/src/main/java/com/emc/ideaforce/service/UserService.java
@@ -67,7 +67,6 @@ public class UserService implements UserDetailsService {
         user.setLastName(userDtoToRegister.getLastName());
         user.setPassword(passwordEncoder.encode(userDtoToRegister.getPassword()));
         user.setEmail(userDtoToRegister.getEmail().toLowerCase());
-        user.setRoles(userDtoToRegister.getRoles());
         User registered = repository.save(user);
         logger.info("Created user {}", user.getEmail());
