+++ ideaforce/src/main/java/com/emc/ideaforce/utils/Utils.java
@@ -3,7 +3,7 @@ package com.emc.ideaforce.utils;
 public class Utils {
     public static final String PWD_PRIVELEGE = "CHANGE_PASSWORD_PRIVILEGE";
     public static final String REG_USER_ROLE = "ROLE_USER";
-    public static final String ADMIN_ROLE = "ADMIN";
+    public static final String ADMIN_ROLE = "ROLE_ADMIN";
 
     public static final String ADMIN = "admin";
 }
