+++ ideaforce/src/main/resources/templates/admin.html
@@ -8,7 +8,7 @@
 </head>
 <body>
 <header>
-    <div th:replace="common/header :: header('admin', userIsAdmin)"></div>
+    <div th:replace="common/header :: header('admin')"></div>
 </header>
 <div class="container main">
     <h2 class="text-center">Admin Page</h2>
