+++ ideaforce/src/main/resources/templates/challengedetail.html
@@ -8,7 +8,7 @@
 </head>
 <body>
 <header>
-    <div th:replace="common/header :: header('challenges', userIsAdmin)"></div>
+    <div th:replace="common/header :: header('challenges')"></div>
 </header>
 <div class="container main">
