+++ ideaforce/src/main/resources/templates/common/header.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
+<html xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
 <head th:fragment="common-header">
 
     <!-- Required meta tags -->
@@ -17,7 +17,7 @@
           integrity="sha384-B4dIYHKNBt8Bc12p+WXckhzcICo0wtJAoU8YZTY5qE0Id1GSseTk6S+L3BlXeVIU" crossorigin="anonymous">
 </head>
 <body>
-<div th:fragment="header(activeTab, userIsAdmin)">
+<div th:fragment="header(activeTab)">
     <div class="container main_banner fixed-top">
         <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
             <a class="navbar-brand" href="/">IdeaForce</a>
@@ -43,7 +43,7 @@
                     <li class="nav-item">
                         <a class="nav-link" href="#">Green Blog</a>
                     </li>
-                    <li th:class="${activeTab == 'admin'}? 'active' : null" th:if="${userIsAdmin}">
+                    <li sec:authorize="hasRole('ROLE_ADMIN')" th:class="${activeTab == 'admin'}? 'active' : null">
                         <a class="nav-link" href="/admin">Pending Challenges</a>
                     </li>
                     <li class="nav-item justify-content-end">
