+++ ideaforce/src/main/resources/templates/gallery.html
@@ -8,7 +8,7 @@
 </head>
 <body>
 <header>
-    <div th:replace="common/header :: header('gallery', userIsAdmin)"></div>
+    <div th:replace="common/header :: header('gallery')"></div>
 </header>
 <div class="container main">
     <h2 class="text-center">Gallery Page</h2>
