+++ ideaforce/src/main/resources/templates/index.html
@@ -8,7 +8,7 @@
 </head>
 <body>
 <header>
-    <div th:replace="common/header :: header('index', userIsAdmin)"></div>
+    <div th:replace="common/header :: header('index')"></div>
 </header>
 <div class="container main">
     <div class="container text-center">
