+++ ideaforce/src/main/resources/templates/leaderboard.html
@@ -8,7 +8,7 @@
 </head>
 <body>
 <header>
-    <div th:replace="common/header :: header('leaderboard', userIsAdmin)"></div>
+    <div th:replace="common/header :: header('leaderboard')"></div>
 </header>
 <div class="container main">
     <h2 class="text-center">LeaderBoard Page</h2>
